avoid re-using stale state in parallel auth flows
Separates transaction creation and resumption into explicitly separate
functions that are called explicitly from the appropriate request
handler in the provider, rather than create/resume being called in order
in the base handler itself.

There are no use-cases where both creation and resumption are expected
to occur in a single request, so the previous implementation is
bug-prone.

Fixes: #9